feat(maps): added map header format
The map header allows a LOT more information to be stored about the
maps, allowing the maps to be far more complex and interesting. For
example, it is now fairly eays to add custom images to map tiles.

Also included a change to the RNG to make it more correct and random.