feat(net): improved sync with multiplayer
Not onyl was respawning added to the game, making it much easier to
actually make games continue for a little bit, but networking was also
changed to in theory support 2+ players.

I also made it so that the RNG seed is set, so that when players shoot
they should have an equivalent experience. Another way of making that
work would be by having players request the RNG seed from the server so
they can just permanently stay in sync.

Players also now send their health to make sure all players see the same
amount of health.