feat(game): added main menu
Added a very simple main menu to the beginning of the game. You still
need to give some input from the terminal, but that should be changing
very soon.

This required a pretty large restructuring of how the code works, to
keep the menu system from becoming too complex in the future. THe code
for different "sections" of the game are split up into different
functions. This currently includes the main menu, the (empty) settings
menu, and the actual game. When you move between states of the game,
you're expected to use the enter_state function.

Also made the map more complex and larger.