feat(game): added the sniper rifle
The sniper rifle shoots very fast projectiles that instakil, but in
exchange the sniper player's position is constantly revealed by a giant
red laser sight, which also helps the sniper aim.

The more interesting engineering side of adding this is that sniper
projectiles move much too fast for the simple aabb collision used for
players and other slower projectiles, so I added a new function that
checks if the projectile hit any object on it's movement from point a to
point b, so that it can't just "jump" over map objects and players. I
expect to eventually use this function for most movement collision
detection in the game. It is unfortuantly much slower, but this hasn't
been an issue as of yet.