tidy(game_mode): cleaned up the game mode code
Moved all of the boilerplate memory related stuff to a file called mem.c
(or mem.rs in the case of the Rust version). This lets people adding a
new game mode to stop worrying about all the boilerplate and just
focus on the game mode.

Also made it much more difficult to screw up the teams pointer thing,
and removed a realloc(????????) in src/net.c which for the life of me I
can't understand how that survived until now.